refactor(machine): unify maxCallDepth/callDepth as int
Stage 1 of Finding 4 from plans/2026-05-06-machine-structural-reduction.md.

Eliminates the hot-path uint64(p.callDepth) cast at
machine_context_continuation.go SaveContinuation by unifying
MachineContext.maxCallDepth (uint64) with vmState.callDepth (int)
on the int side. callDepth must stay signed because the underflow
guard in PopContinuation (`p.callDepth--; if p.callDepth < 0`)
relies on signed-comparison semantics; uint64 would wrap to
MaxUint64 and silently turn a compiler-bug diagnostic into a
runaway.

Public API changes (Wile v1.x, zero consumers — break-freely
policy per CLAUDE.md):
  - WithMaxCallDepth(n uint64) -> WithMaxCallDepth(n int)
  - DefaultMaxCallDepth uint64 = 10000 -> int = 10000
  - MachineContext.MaxCallDepth() uint64 -> int
  - MachineContext.SetMaxCallDepth(uint64) -> SetMaxCallDepth(int)
  - SubContextParams.MaxCallDepth uint64 -> int

Sentinel "0 = unlimited" semantics preserved — see Stage 2 (a
separate commit, bench-gated) for the sentinel-to-nullable
conversion proposed in the plan.

Tests: machine_context_test.go and wile_test.go field/literal
types updated to int. All other tests unchanged. make lint and
make covercheck pass.

+++ engine.go
@@ -42,7 +42,7 @@ var ErrEngineClosed = werr.NewStaticError("engine is closed")
 // DefaultMaxCallDepth is the default call depth limit for new engines.
 // At ~500 bytes per frame, 10000 frames ≈ 5MB. Use WithMaxCallDepth(0)
 // to opt out of the limit explicitly.
-const DefaultMaxCallDepth uint64 = 10000
+const DefaultMaxCallDepth int = 10000
 
 // Engine is the main entry point for embedding Wile.
 //
@@ -61,7 +61,7 @@ type Engine struct {
 	lastCounters        machine.VMCounters
 	closers             []registry.Closeable
 	closed              bool
-	maxCallDepth        uint64
+	maxCallDepth        int
 	maxStackSize        uint64
 	inlineThreshold     int
 	contractEnforcement bool // propagated to RegisterPrimitive via cfg
